Wireless mouse
Wireless Mouse require WPC approval, must go through a certification process. The first step is to submit an application along with the required documents including RF Test report to the WPC office. Once the application is reviewed and approved, the applicant will be issued a certificate – ETA Approval. The product can then be marketed and sold in India.